correct a bug parsing the applications rating add unique name to applications

+++ applications/src/main/java/com/asfoundation/wallet/apps/repository/webservice/data/Rating.java
@@ -5,10 +5,10 @@ import com.google.gson.annotations.SerializedName;
 
 public class Rating {
 
-  @SerializedName("avg") @Expose private int avg;
-  @SerializedName("total") @Expose private int total;
+  @SerializedName("avg") @Expose private float avg;
+  @SerializedName("total") @Expose private float total;
 
-  public int getAvg() {
+  public float getAvg() {
     return avg;
   }
 
@@ -16,7 +16,7 @@ public class Rating {
     this.avg = avg;
   }
 
-  public int getTotal() {
+  public float getTotal() {
     return total;
   }
